+++ nvim/lua/yvhn/init.lua
@@ -1,6 +1,7 @@
 require("yvhn.settings")
 require("yvhn.persist")
 require("yvhn.theme")
+require("yvhn.utils")
 require("yvhn.remaps")
 require("yvhn.lsp")
 require("yvhn.treesitter_context")
